requirements.txt
presidio-analyzer
presidio-anonymizer
presidio-image-redactor
spacy
pymupdf
pytesseract
pillow
pyyaml
